Split the composer context object out of the provider
Reading the composer context meant importing the provider, and the
provider imports the wallet, which imports webext-bridge, which calls
chrome.runtime.connect the moment the module loads. Under jsdom that
throws before any test runs, so two component tests had to mock the
whole context module to render a screen that wanted three fields off
the decoded message. The mock is not free either: it stubs the hook the
component actually uses, so the test stops covering it.

The context object, its types and the read hooks now live in a leaf
module whose every import is type-only and therefore erased, leaving
React as the only thing that follows it at runtime. The provider
imports the object from there. Consumers import the hooks directly
rather than through a re-export, so the dependency stays legible.

One of the two mocks is gone entirely and the other now stubs only the
hook, not the module that defines it.

Comments in the moved code were rewritten rather than carried over.
Most of them restated the field name or its type -- "Error message to
display" above `error: string | null` -- and uniform annotation is
worse than none, because the few comments that carry real information
get skimmed along with the rest. What survives is the reason the file
is split, the fee rate's unit and null meaning, where the decoded
message comes from, and what goBack actually does.

Some invented shorthand went with it: "glue", "harness" and
"principled declines" all compressed an idea into a label that only
made sense to whoever coined it.
