+++ src/components/composer/composer.tsx
@@ -4,7 +4,8 @@ import { useNavigate } from "react-router";
 import { SuccessScreen } from "@/components/screens/success-screen";
 import { Spinner } from "@/components/ui/spinner";
 import { Banner } from "@/components/ui/banner";
-import { ComposerProvider, useComposer } from "@/contexts/composer-context";
+import { ComposerProvider } from "@/contexts/composer-context"
+import { useComposer } from "@/contexts/composer-context-object";
 import { useHeader } from "@/contexts/header-context";
 import type { ApiResponse } from "@/utils/blockchain/counterparty/compose";
