+++ src/pages/compose/broadcast/address-options/form.tsx
@@ -4,7 +4,7 @@ import { Field, Label } from "@headlessui/react";
 import { ComposerForm } from "@/components/composer/composer-form";
 import { AddressHeader } from "@/components/ui/headers/address-header";
 import { CheckboxInput } from "@/components/ui/inputs/checkbox-input";
-import { useComposer } from "@/contexts/composer-context";
+import { useComposer } from "@/contexts/composer-context-object";
 import type { BroadcastOptions } from "@/utils/blockchain/counterparty/compose";
 import type { ReactElement } from "react";
