+++ src/pages/compose/dispenser/close-by-hash/form.tsx
@@ -3,7 +3,7 @@ import { useFormStatus } from "react-dom";
 import { ComposerForm } from "@/components/composer/composer-form";
 import { HashInput } from "@/components/ui/inputs/hash-input";
 import { AddressHeader } from "@/components/ui/headers/address-header";
-import { useComposer } from "@/contexts/composer-context";
+import { useComposer } from "@/contexts/composer-context-object";
 import { fetchDispenserByHash } from "@/utils/blockchain/counterparty/api";
 import type { DispenserOptions } from "@/utils/blockchain/counterparty/compose";
 import type { ReactElement } from "react";
