+++ src/pages/compose/issuance/update-description/form.tsx
@@ -5,7 +5,7 @@ import { ComposerForm } from "@/components/composer/composer-form";
 import { Spinner } from "@/components/ui/spinner";
 import { SettingSwitch } from "@/components/ui/inputs/setting-switch";
 import { InscriptionUploadInput } from "@/components/ui/inputs/file-upload-input";
-import { useComposer } from "@/contexts/composer-context";
+import { useComposer } from "@/contexts/composer-context-object";
 import { useAssetInfo } from "@/hooks/useAssetInfo";
 import { isSegwitFormat } from '@/utils/blockchain/bitcoin/address';
 import type { IssuanceOptions } from "@/utils/blockchain/counterparty/compose";
