+++ src/pages/compose/order/btcpay/form.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import { ComposerForm } from "@/components/composer/composer-form";
 import { HashInput } from "@/components/ui/inputs/hash-input";
 import { AddressHeader } from "@/components/ui/headers/address-header";
-import { useComposer } from "@/contexts/composer-context";
+import { useComposer } from "@/contexts/composer-context-object";
 import type { BTCPayOptions } from "@/utils/blockchain/counterparty/compose";
 import type { ReactElement } from "react";
