+++ src/pages/compose/send/mpma/form.tsx
@@ -1,6 +1,6 @@
 import { useState, useRef } from "react";
 import { ComposerForm } from "@/components/composer/composer-form";
-import { useComposer } from "@/contexts/composer-context";
+import { useComposer } from "@/contexts/composer-context-object";
 import { fetchAssetDetails } from "@/utils/blockchain/counterparty/api";
 import { isHexMemo, stripHexPrefix, isValidMemoLength } from "@/utils/blockchain/counterparty/memo";
 import { validateBitcoinAddress } from "@/utils/validation/bitcoin";
